Slow and precise movement

+++ src/com/adamkacik/game/entity/Entity.java
@@ -8,7 +8,7 @@ import com.adamkacik.game.level.Level;
 
 public class Entity {
 	
-		protected int x,y;
+		protected double x,y;
 		protected Sprite sprite;
 		public boolean removed = false;
 		protected Level level;
@@ -27,12 +27,12 @@ public class Entity {
 		}
 		
 		public void render(Screen screen) {
-			if (sprite != null) screen.renderSprite(x, y, sprite, true);
+			if (sprite != null) screen.renderSprite((int)x, (int)y, sprite, true);
 		}
-		public int getX() {
+		public double getX() {
 			return x;
 		}
-		public int getY() {
+		public double getY() {
 			return y;
 		}
 		public void remove() {
